Adding coments, refactoring code.

+++ Battleships/Models/BattleshipJsonObject.cs
@@ -1,15 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Battleships.Models
 {
     public class BattleshipJsonObject
     {
         public string UserName { get; set; }
-        public int PlayedTime { get; set; }
+        public double PlayedTime { get; set; }
         public int Score { get; set; }
         public Guid Id { get; set; }
     }
